+++ statham/dsl/parser.py
@@ -154,14 +154,19 @@ def _new_object(
         title = f"{title}_{count}"
     default = schema.get("default", NotPassed())
     required = set(schema.get("required", []))
+    attr_name = lambda key: key if key != "default" else f"_{key}"
     properties = {
         # TODO: Handle attribute names which don't work in python.
-        key: _Property(parse_element(value, counter), required=key in required)
+        attr_name(key): _Property(
+            parse_element(value, counter), required=key in required, name=key
+        )
         for key, value in schema.get("properties", {}).items()
         # Ignore malformed values.
         if isinstance(value, dict)
     }
-    class_dict = ObjectClassDict(default=default, **properties)
+    class_dict = ObjectClassDict(default=default)
+    for key, value in properties.items():
+        class_dict[key] = value
     return ObjectMeta(_title_format(title), (Object,), class_dict)