+++ tests/test_serializer.py
@@ -2,7 +2,7 @@ from statham.dsl.constants import Maybe
 from statham.dsl.elements import Object, String
 from statham.dsl.parser import parse
 from statham.dsl.property import Property
-from statham.serializer import serialize_python, _serialize_object
+from statham.serializer import _IMPORT_STATEMENTS, serialize_python
 
 
 SCHEMA = {
@@ -15,7 +15,8 @@ SCHEMA = {
             "title": "Category",
             "properties": {"value": {"type": "string"}},
             "default": {"value": "none"},
-        }
+        },
+        "default": {"type": "string"},
     },
     "definitions": {
         "other": {
@@ -28,26 +29,8 @@ SCHEMA = {
 
 
 def test_schema_reserializes_to_expected_python_string():
-    assert (
-        serialize_python(*parse(SCHEMA))
-        == """from typing import List, Union
-
-from statham.dsl.constants import Maybe
-from statham.dsl.elements import (
-    AnyOf,
-    Array,
-    Boolean,
-    Integer,
-    Null,
-    Number,
-    OneOf,
-    Object,
-    String,
-)
-from statham.dsl.property import Property
-
-
-class Category(Object):
+    assert serialize_python(*parse(SCHEMA)) == _IMPORT_STATEMENTS + (
+        """class Category(Object):
 
     default = {'value': 'none'}
 
@@ -58,10 +41,36 @@ class Parent(Object):
 
     category: Category = Property(Category, required=True)
 
+    _default: Maybe[str] = Property(String(), name='default')
+
 
 class Other(Object):
 
     value: Maybe[int] = Property(Integer())
+"""
+    )
+
+
+def test_parse_and_serialize_schema_with_self_property():
+    schema = {
+        "type": "object",
+        "title": "MyObject",
+        "properties": {"self": {"type": "string"}},
+    }
+    assert serialize_python(*parse(schema)) == _IMPORT_STATEMENTS + (
+        """class MyObject(Object):
+
+    self: Maybe[str] = Property(String())
+"""
+    )
+
+
+def test_parse_and_serialize_schema_with_no_args():
+    schema = {"type": "object", "title": "NoProps"}
+    assert serialize_python(*parse(schema)) == _IMPORT_STATEMENTS + (
+        """class NoProps(Object):
+
+    pass
 """
     )
 
@@ -83,7 +92,7 @@ class ObjectWrapper(Object):
 
 def test_serialize_string_wrapper_object():
     assert (
-        _serialize_object(StringWrapper)
+        StringWrapper.python()
         == """class StringWrapper(Object):
 
     value: str = Property(String(), required=True)
@@ -93,7 +102,7 @@ def test_serialize_string_wrapper_object():
 
 def test_serialize_object_wrapper_object():
     assert (
-        _serialize_object(ObjectWrapper)
+        ObjectWrapper.python()
         == """class ObjectWrapper(Object):
 
     value: Maybe[StringWrapper] = Property(StringWrapper)