+++ CHANGELOG.md
@@ -13,6 +13,12 @@ Types of changes are:
 * **Fixed** for any bug fixes.
 
 ## [Unreleased]
+### Added
+* Added support for `default` as a property name in object schemas.
+* Added support for `self` as a property name in object schemas.
+
+### Changed
+* Moved bulk to python serialization logic to DSL element methods.
 
 ## [0.2.0] - 2020-03-20
 ### Added